Added local test for Charge list request and deleted old resource test
