+++ src/test/java/co/omise/requests/ChargeRequestTest.java
@@ -2,7 +2,9 @@ package co.omise.requests;
 
 import co.omise.models.Charge;
 import co.omise.models.OmiseException;
+import co.omise.models.ScopedList;
 import co.omise.models.SourceType;
+import com.fasterxml.jackson.core.type.TypeReference;
 import org.junit.Test;
 
 import java.io.IOException;
@@ -100,4 +102,21 @@ public class ChargeRequestTest extends RequestTest {
         assertEquals(CHARGE_ID, charge.getId());
         assertTrue(charge.isReversed());
     }
+
+    @Test
+    public void testList() throws IOException, OmiseException {
+        Request<Charge> listChargeRequest =
+                new Charge.ListRequestBuilder().build();
+        ScopedList<Charge> list = getTestRequester().sendRequest(listChargeRequest, new TypeReference<ScopedList<Charge>>() {
+        });
+
+        assertRequested("GET", "/charges", 200);
+        assertEquals(2, list.getTotal());
+        assertEquals(20, list.getLimit());
+
+        Charge charge = list.getData().get(0);
+        assertEquals("charge", charge.getObject());
+        assertEquals(CHARGE_ID, charge.getId());
+    }
+
 }
